requirements-dev.txt
-r requirements.txt
codecov
flake8
pysimplemodel
pytest
pytest-cov
Sphinx
twine
wheel
